Promote dummy key to actual key class
